Move things into already existing packages
