+++ src/main/kotlin/App.kt
@@ -1,6 +1,5 @@
 import controllers.CacheChooserController
-import environmentexporter.CliExporter
-import environmentexporter.StartupOptions
+import models.StartupOptions
 import models.config.ConfigOptions
 import models.config.Configuration
 import org.pushingpixels.radiance.theming.api.skin.RadianceGraphiteAquaLookAndFeel
